+++ mkosi/distributions/opensuse.py
@@ -13,8 +13,9 @@ from mkosi.installer.dnf import Dnf
 from mkosi.installer.rpm import RpmRepository, find_rpm_gpgkey, setup_rpm
 from mkosi.installer.zypper import Zypper
 from mkosi.log import die
+from mkosi.mounts import finalize_crypto_mounts
 from mkosi.run import find_binary, run
-from mkosi.sandbox import Mount, finalize_crypto_mounts
+from mkosi.sandbox import Mount
 from mkosi.util import listify, sort_packages
 
 
@@ -165,6 +166,7 @@ def fetch_gpgurls(context: Context, repourl: str) -> tuple[str, ...]:
                 "--no-progress-meter",
                 "--fail",
                 *(["--proxy", context.config.proxy_url] if context.config.proxy_url else []),
+                *(["--noproxy", ",".join(context.config.proxy_exclude)] if context.config.proxy_exclude else []),
                 *(["--proxy-capath", "/proxy.cacert"] if context.config.proxy_peer_certificate else []),
                 *(["--proxy-cert", "/proxy.clientcert"] if context.config.proxy_client_certificate else []),
                 *(["--proxy-key", "/proxy.clientkey"] if context.config.proxy_client_key else []),
@@ -172,7 +174,7 @@ def fetch_gpgurls(context: Context, repourl: str) -> tuple[str, ...]:
             ],
             sandbox=context.sandbox(
                 network=True,
-                mounts=[Mount(d, d), *finalize_crypto_mounts(context.config.tools())],
+                mounts=[Mount(d, d), *finalize_crypto_mounts(context.config)],
             ),
         )
         xml = (Path(d) / "repomd.xml").read_text()