+++ mkosi/run.py
@@ -15,6 +15,7 @@ import subprocess
 import sys
 import threading
 from collections.abc import Awaitable, Collection, Iterator, Mapping, Sequence
+from contextlib import AbstractContextManager
 from pathlib import Path
 from types import TracebackType
 from typing import Any, Callable, NoReturn, Optional
@@ -134,9 +135,10 @@ def run(
     env: Mapping[str, str] = {},
     cwd: Optional[Path] = None,
     log: bool = True,
-    preexec_fn: Optional[Callable[[], None]] = None,
-    sandbox: Sequence[PathString] = (),
     foreground: bool = True,
+    preexec_fn: Optional[Callable[[], None]] = None,
+    success_exit_status: Sequence[int] = (0,),
+    sandbox: AbstractContextManager[Sequence[PathString]] = contextlib.nullcontext([]),
 ) -> CompletedProcess:
     if input is not None:
         assert stdin is None  # stdin and input cannot be specified together
@@ -153,10 +155,12 @@ def run(
         env=env,
         cwd=cwd,
         log=log,
-        preexec_fn=preexec_fn,
-        sandbox=sandbox,
         foreground=foreground,
-    ) as process:
+        preexec_fn=preexec_fn,
+        success_exit_status=success_exit_status,
+        sandbox=sandbox,
+        innerpid=False,
+    ) as (process, _):
         out, err = process.communicate(input)
 
     return CompletedProcess(cmdline, process.returncode, out, err)
@@ -177,11 +181,12 @@ def spawn(
     log: bool = True,
     foreground: bool = False,
     preexec_fn: Optional[Callable[[], None]] = None,
-    sandbox: Sequence[PathString] = (),
-) -> Iterator[Popen]:
+    success_exit_status: Sequence[int] = (0,),
+    sandbox: AbstractContextManager[Sequence[PathString]] = contextlib.nullcontext([]),
+    innerpid: bool = True,
+) -> Iterator[tuple[Popen, int]]:
     assert sorted(set(pass_fds)) == list(pass_fds)
 
-    sandbox = [os.fspath(x) for x in sandbox]
     cmdline = [os.fspath(x) for x in cmdline]
 
     if ARG_DEBUG.get():
@@ -244,72 +249,98 @@ def spawn(
             # expect it to pick.
             assert nfd == SD_LISTEN_FDS_START + i
 
-    # First, check if the sandbox works at all before executing the command.
-    if sandbox and (rc := subprocess.run(sandbox + ["true"]).returncode) != 0:
-        log_process_failure(sandbox, cmdline, rc)
-        raise subprocess.CalledProcessError(rc, sandbox + cmdline)
+    with sandbox as sbx:
+        prefix = [os.fspath(x) for x in sbx]
 
-    if subprocess.run(sandbox + ["sh", "-c", f"command -v {cmdline[0]}"], stdout=subprocess.DEVNULL).returncode != 0:
-        die(f"{cmdline[0]} not found.", hint=f"Is {cmdline[0]} installed on the host system?")
+        # First, check if the sandbox works at all before executing the command.
+        if prefix and (rc := subprocess.run(prefix + ["true"]).returncode) != 0:
+            log_process_failure(prefix, cmdline, rc)
+            raise subprocess.CalledProcessError(rc, prefix + cmdline)
 
-    if (
-        foreground and
-        sandbox and
-        subprocess.run(sandbox + ["sh", "-c", "command -v setpgid"], stdout=subprocess.DEVNULL).returncode == 0
-    ):
-        sandbox += ["setpgid", "--foreground", "--"]
+        if subprocess.run(
+            prefix + ["sh", "-c", f"command -v {cmdline[0]}"],
+            stdout=subprocess.DEVNULL,
+        ).returncode != 0:
+            die(f"{cmdline[0]} not found.", hint=f"Is {cmdline[0]} installed on the host system?")
 
-    if pass_fds:
-        # We don't know the PID before we start the process and we can't modify the environment in preexec_fn so we
-        # have to spawn a temporary shell to set the necessary environment variables before spawning the actual
-        # command.
-        sandbox += ["sh", "-c", f"LISTEN_FDS={len(pass_fds)} LISTEN_PID=$$ exec $0 \"$@\""]
+        if (
+            foreground and
+            prefix and
+            subprocess.run(prefix + ["sh", "-c", "command -v setpgid"], stdout=subprocess.DEVNULL).returncode == 0
+        ):
+            prefix += ["setpgid", "--foreground", "--"]
 
-    try:
-        with subprocess.Popen(
-            sandbox + cmdline,
-            stdin=stdin,
-            stdout=stdout,
-            stderr=stderr,
-            text=True,
-            user=user,
-            group=group,
-            # pass_fds only comes into effect after python has invoked the preexec function, so we make sure that
-            # pass_fds contains the file descriptors to keep open after we've done our transformation in preexec().
-            pass_fds=[SD_LISTEN_FDS_START + i for i in range(len(pass_fds))],
-            env=env,
-            cwd=cwd,
-            preexec_fn=preexec,
-        ) as proc:
-            try:
-                yield proc
-            except BaseException:
-                proc.terminate()
-                raise
-            finally:
-                returncode = proc.wait()
+        if pass_fds:
+            # We don't know the PID before we start the process and we can't modify the environment in preexec_fn so we
+            # have to spawn a temporary shell to set the necessary environment variables before spawning the actual
+            # command.
+            prefix += ["sh", "-c", f"LISTEN_FDS={len(pass_fds)} LISTEN_PID=$$ exec $0 \"$@\""]
 
-            if check and returncode:
-                if log:
-                    log_process_failure(sandbox, cmdline, returncode)
-                if ARG_DEBUG_SHELL.get():
-                    subprocess.run(
-                        [*sandbox, "bash"],
-                        check=False,
-                        stdin=sys.stdin,
-                        text=True,
-                        user=user,
-                        group=group,
-                        env=env,
-                        cwd=cwd,
-                        preexec_fn=preexec,
-                    )
-                raise subprocess.CalledProcessError(returncode, cmdline)
-    except FileNotFoundError as e:
-        die(f"{e.filename} not found.")
-    finally:
-        if foreground:
-            make_foreground_process(new_process_group=False)
+        if prefix and innerpid:
+            r, w = os.pipe2(os.O_CLOEXEC)
+            # Make sure that the write end won't be overridden in preexec() when we're moving fds forward.
+            q = fcntl.fcntl(w, fcntl.F_DUPFD_CLOEXEC, SD_LISTEN_FDS_START + len(pass_fds) + 1)
+            os.close(w)
+            w = q
+            # dash doesn't support working with file descriptors higher than 9 so make sure we use bash.
+            prefix += ["bash", "-c", f"echo $$ >&{w} && exec {w}>&- && exec $0 \"$@\""]
+        else:
+            r, w = (None, None)
+
+        try:
+            with subprocess.Popen(
+                prefix + cmdline,
+                stdin=stdin,
+                stdout=stdout,
+                stderr=stderr,
+                text=True,
+                user=user,
+                group=group,
+                # pass_fds only comes into effect after python has invoked the preexec function, so we make sure that
+                # pass_fds contains the file descriptors to keep open after we've done our transformation in preexec().
+                pass_fds=[SD_LISTEN_FDS_START + i for i in range(len(pass_fds))] + ([w] if w else []),
+                env=env,
+                cwd=cwd,
+                preexec_fn=preexec,
+            ) as proc:
+                if w:
+                    os.close(w)
+                pid = proc.pid
+                try:
+                    if r:
+                        with open(r) as f:
+                            s = f.read()
+                            if s:
+                                pid = int(s)
+
+                    yield proc, pid
+                except BaseException:
+                    kill(proc, pid, signal.SIGTERM)
+                    raise
+                finally:
+                    returncode = proc.wait()
+
+                if check and returncode not in success_exit_status:
+                    if log:
+                        log_process_failure(prefix, cmdline, returncode)
+                    if ARG_DEBUG_SHELL.get():
+                        subprocess.run(
+                            [*prefix, "bash"],
+                            check=False,
+                            stdin=sys.stdin,
+                            text=True,
+                            user=user,
+                            group=group,
+                            env=env,
+                            cwd=cwd,
+                            preexec_fn=preexec,
+                        )
+                    raise subprocess.CalledProcessError(returncode, cmdline)
+        except FileNotFoundError as e:
+            die(f"{e.filename} not found.")
+        finally:
+            if foreground:
+                make_foreground_process(new_process_group=False)
 
 
 def find_binary(*names: PathString, root: Path = Path("/")) -> Optional[Path]:
@@ -331,6 +362,18 @@ def find_binary(*names: PathString, root: Path = Path("/")) -> Optional[Path]:
                 return Path("/") / Path(binary).relative_to(root)
 
     return None
+
+
+def kill(process: Popen, innerpid: int, signal: int) -> None:
+    process.poll()
+    if process.returncode is not None:
+        return
+
+    try:
+        os.kill(innerpid, signal)
+    # Handle the race condition where the process might exit between us calling poll() and us calling os.kill().
+    except ProcessLookupError:
+        pass
 
 
 class AsyncioThread(threading.Thread):